requirements.txt
streamlit==1.52.1
google-generativeai==0.8.6
